Ajuste arquitetura usuario
