+++ src/routes/UserRouter.ts
@@ -1,3 +1,9 @@
-import express from 'express';
+import express from "express";
+import { signup } from "../controllers/user/signup";
+import { login } from "../controllers/user/login";
 
 export const userRouter = express.Router();
+
+userRouter.post("/signup", signup);
+
+userRouter.post("/login", login);
